Rewrite for julia 0.7
The strategy used here for ColorView and ChannelView has been
implemented in a much more general way for ReinterpretArray. It makes
sense to leverage the Base implementation. Plus, now that
`reinterpret` is no longer restricted to `Array`, it turns out that we
run into certain dispatch problems that stem from changing the
dimensionality to handle/subsume color channels. The solution here is
to introduce a `reinterpretc` that is specifically for handling
colors.

This definitely breaks old code.