Switch to SnoopPrecompile
Using Julia nightly, this cuts the time required to execute the
workload in the `@precompile_all_calls` from 6.55s to 2.15s,
while also slightly reducing load time from 1.06s to 0.99s.
The fact that it's not even larger is surely due to compilation
of the helper functions `pcarray`, `pcm`, and `pcmv`.